+++ src/StructuredField.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
-interface StructuredField
+use Stringable;
+
+interface StructuredField extends Stringable
 {
     /**
      * Returns the serialize-representation of the Structured Field as a textual HTTP field value.